refactor: common domain package 이동
