+++ src/main/java/kitchenpos/common/domain/Price.java
@@ -1,4 +1,4 @@
-package kitchenpos.domain.common;
+package kitchenpos.common.domain;
 
 import java.math.BigDecimal;
 import java.util.Objects;
