+++ src/main/java/kitchenpos/domain/MenuGroup.java
@@ -6,7 +6,7 @@ import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
 import javax.persistence.Id;
 import javax.persistence.Table;
-import kitchenpos.domain.common.Name;
+import kitchenpos.common.domain.Name;
 
 @Entity
 @Table(name = "menu_group")
