+++ src/main/java/kitchenpos/domain/MenuProduct.java
@@ -9,8 +9,8 @@ import javax.persistence.Id;
 import javax.persistence.JoinColumn;
 import javax.persistence.ManyToOne;
 import javax.persistence.Table;
-import kitchenpos.domain.common.Price;
-import kitchenpos.domain.common.Quantity;
+import kitchenpos.common.domain.Price;
+import kitchenpos.common.domain.Quantity;
 
 @Entity
 @Table(name = "menu_product")
