+++ src/test/java/kitchenpos/domain/common/NameTest.java
@@ -2,6 +2,7 @@ package kitchenpos.domain.common;
 
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
+import kitchenpos.common.domain.Name;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
