+++ src/test/java/kitchenpos/domain/common/QuantityTest.java
@@ -3,6 +3,7 @@ package kitchenpos.domain.common;
 
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
+import kitchenpos.common.domain.Quantity;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
